2010-06-19 13:41 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * include/hbapi.h   * include/hbapiitm.h   * src/vm/itemapi.c   * src/vm/arrays.c   * src/vm/extend.c     + Added hb_storvns().     + Added hb_parvns().     ; TOFIX: hb_itemPutNS() gives these (not quite unexpected) mingw warnings:              ---              itemapi.c:993:8: warning: comparison between signed and unsigned integer expressions              itemapi.c:998:38: warning: comparison between signed and unsigned integer expressions              itemapi.c:1004:35: warning: comparison between signed and unsigned integer expressions              itemapi.c:1004:7: warning: comparison is always false due to limited range of data type              ---
  * src/rtl/philes.c
  * src/rtl/mlcfunc.c
  * src/rtl/rat.c
    + Using size APIs.
    % Also deleting '( HB_SIZE )' casts along the way.

  * src/rtl/oemansi.c
    ! Cast to make Win64 happy. Quite interesting that almost no
      Windows API don't support proper 64-bit strings, so here
      Windows will be the limitation, not Harbour.

+++ harbour/src/rtl/oemansi.c
@@ -67,7 +67,7 @@ HB_FUNC( HB_ANSITOOEM )
    if( pString )
 #if defined( HB_OS_WIN )
    {
-      int nLen = hb_itemGetCLen( pString );
+      int nLen = ( int ) hb_itemGetCLen( pString );
       const char * pszSrc = hb_itemGetCPtr( pString );
 
       int nWideLen = MultiByteToWideChar( CP_ACP, MB_PRECOMPOSED, pszSrc, nLen, NULL, 0 );
@@ -99,7 +99,7 @@ HB_FUNC( HB_OEMTOANSI )
    if( pString )
 #if defined( HB_OS_WIN )
    {
-      int nLen = hb_itemGetCLen( pString );
+      int nLen = ( int ) hb_itemGetCLen( pString );
       const char * pszSrc = hb_itemGetCPtr( pString );
 
       int nWideLen = MultiByteToWideChar( CP_OEMCP, MB_PRECOMPOSED, pszSrc, nLen, NULL, 0 );
